Aesthetic mappings for datasets

The idea comes straight from aes_ in ggplot2. That provides a way to map columns of a dataset to features of graphic. Here we are mapping columns into a function so we can use standard names inside that function. This function variant expects string argument, rather than the unquoted version aes_d

aes_d_(...)

Arguments

...

Unquoted, comma-seperated column mappings

Details

A mapping looks like: <column_to_be_created> = <existing column>

Examples

aes_d(group = 'class')
#> $group
#> [1] "class"
#>