Waffle Iron

Pour your data into a waffle iron to get ready to cook a waffle chart.

waffle_iron(data, mapping, rows = 8, sample = 1)

Arguments

data
A dataframe to feed into the waffle iron

mapping
A mapping as produce by aes_d

rows
The number of rows in the waffle

sample
The proportion of rows to sample the dataset (between 0 and 1). Useful when the dataset is too large to plot correctly.

Details

Prepare raw data so it is fit to create a waffle visualisation. The type of data transformation that is required does not gel well with ggplot2 underlying mechanism. The way around this is to provide a function that does the preperation outside of ggplot.

Examples

waffle_iron(mpg, aes_d(group = class))

#>     y  x      group
#> 1   1  1    2seater
#> 2   2  1    2seater
#> 3   3  1    2seater
#> 4   4  1    2seater
#> 5   5  1    2seater
#> 6   6  1    compact
#> 7   7  1    compact
#> 8   8  1    compact
#> 9   1  2    compact
#> 10  2  2    compact
#> 11  3  2    compact
#> 12  4  2    compact
#> 13  5  2    compact
#> 14  6  2    compact
#> 15  7  2    compact
#> 16  8  2    compact
#> 17  1  3    compact
#> 18  2  3    compact
#> 19  3  3    compact
#> 20  4  3    compact
#> 21  5  3    compact
#> 22  6  3    compact
#> 23  7  3    compact
#> 24  8  3    compact
#> 25  1  4    compact
#> 26  2  4    compact
#> 27  3  4    compact
#> 28  4  4    compact
#> 29  5  4    compact
#> 30  6  4    compact
#> 31  7  4    compact
#> 32  8  4    compact
#> 33  1  5    compact
#> 34  2  5    compact
#> 35  3  5    compact
#> 36  4  5    compact
#> 37  5  5    compact
#> 38  6  5    compact
#> 39  7  5    compact
#> 40  8  5    compact
#> 41  1  6    compact
#> 42  2  6    compact
#> 43  3  6    compact
#> 44  4  6    compact
#> 45  5  6    compact
#> 46  6  6    compact
#> 47  7  6    compact
#> 48  8  6    compact
#> 49  1  7    compact
#> 50  2  7    compact
#> 51  3  7    compact
#> 52  4  7    compact
#> 53  5  7    midsize
#> 54  6  7    midsize
#> 55  7  7    midsize
#> 56  8  7    midsize
#> 57  1  8    midsize
#> 58  2  8    midsize
#> 59  3  8    midsize
#> 60  4  8    midsize
#> 61  5  8    midsize
#> 62  6  8    midsize
#> 63  7  8    midsize
#> 64  8  8    midsize
#> 65  1  9    midsize
#> 66  2  9    midsize
#> 67  3  9    midsize
#> 68  4  9    midsize
#> 69  5  9    midsize
#> 70  6  9    midsize
#> 71  7  9    midsize
#> 72  8  9    midsize
#> 73  1 10    midsize
#> 74  2 10    midsize
#> 75  3 10    midsize
#> 76  4 10    midsize
#> 77  5 10    midsize
#> 78  6 10    midsize
#> 79  7 10    midsize
#> 80  8 10    midsize
#> 81  1 11    midsize
#> 82  2 11    midsize
#> 83  3 11    midsize
#> 84  4 11    midsize
#> 85  5 11    midsize
#> 86  6 11    midsize
#> 87  7 11    midsize
#> 88  8 11    midsize
#> 89  1 12    midsize
#> 90  2 12    midsize
#> 91  3 12    midsize
#> 92  4 12    midsize
#> 93  5 12    midsize
#> 94  6 12    minivan
#> 95  7 12    minivan
#> 96  8 12    minivan
#> 97  1 13    minivan
#> 98  2 13    minivan
#> 99  3 13    minivan
#> 100 4 13    minivan
#> 101 5 13    minivan
#> 102 6 13    minivan
#> 103 7 13    minivan
#> 104 8 13    minivan
#> 105 1 14     pickup
#> 106 2 14     pickup
#> 107 3 14     pickup
#> 108 4 14     pickup
#> 109 5 14     pickup
#> 110 6 14     pickup
#> 111 7 14     pickup
#> 112 8 14     pickup
#> 113 1 15     pickup
#> 114 2 15     pickup
#> 115 3 15     pickup
#> 116 4 15     pickup
#> 117 5 15     pickup
#> 118 6 15     pickup
#> 119 7 15     pickup
#> 120 8 15     pickup
#> 121 1 16     pickup
#> 122 2 16     pickup
#> 123 3 16     pickup
#> 124 4 16     pickup
#> 125 5 16     pickup
#> 126 6 16     pickup
#> 127 7 16     pickup
#> 128 8 16     pickup
#> 129 1 17     pickup
#> 130 2 17     pickup
#> 131 3 17     pickup
#> 132 4 17     pickup
#> 133 5 17     pickup
#> 134 6 17     pickup
#> 135 7 17     pickup
#> 136 8 17     pickup
#> 137 1 18     pickup
#> 138 2 18 subcompact
#> 139 3 18 subcompact
#> 140 4 18 subcompact
#> 141 5 18 subcompact
#> 142 6 18 subcompact
#> 143 7 18 subcompact
#> 144 8 18 subcompact
#> 145 1 19 subcompact
#> 146 2 19 subcompact
#> 147 3 19 subcompact
#> 148 4 19 subcompact
#> 149 5 19 subcompact
#> 150 6 19 subcompact
#> 151 7 19 subcompact
#> 152 8 19 subcompact
#> 153 1 20 subcompact
#> 154 2 20 subcompact
#> 155 3 20 subcompact
#> 156 4 20 subcompact
#> 157 5 20 subcompact
#> 158 6 20 subcompact
#> 159 7 20 subcompact
#> 160 8 20 subcompact
#> 161 1 21 subcompact
#> 162 2 21 subcompact
#> 163 3 21 subcompact
#> 164 4 21 subcompact
#> 165 5 21 subcompact
#> 166 6 21 subcompact
#> 167 7 21 subcompact
#> 168 8 21 subcompact
#> 169 1 22 subcompact
#> 170 2 22 subcompact
#> 171 3 22 subcompact
#> 172 4 22 subcompact
#> 173 5 22        suv
#> 174 6 22        suv
#> 175 7 22        suv
#> 176 8 22        suv
#> 177 1 23        suv
#> 178 2 23        suv
#> 179 3 23        suv
#> 180 4 23        suv
#> 181 5 23        suv
#> 182 6 23        suv
#> 183 7 23        suv
#> 184 8 23        suv
#> 185 1 24        suv
#> 186 2 24        suv
#> 187 3 24        suv
#> 188 4 24        suv
#> 189 5 24        suv
#> 190 6 24        suv
#> 191 7 24        suv
#> 192 8 24        suv
#> 193 1 25        suv
#> 194 2 25        suv
#> 195 3 25        suv
#> 196 4 25        suv
#> 197 5 25        suv
#> 198 6 25        suv
#> 199 7 25        suv
#> 200 8 25        suv
#> 201 1 26        suv
#> 202 2 26        suv
#> 203 3 26        suv
#> 204 4 26        suv
#> 205 5 26        suv
#> 206 6 26        suv
#> 207 7 26        suv
#> 208 8 26        suv
#> 209 1 27        suv
#> 210 2 27        suv
#> 211 3 27        suv
#> 212 4 27        suv
#> 213 5 27        suv
#> 214 6 27        suv
#> 215 7 27        suv
#> 216 8 27        suv
#> 217 1 28        suv
#> 218 2 28        suv
#> 219 3 28        suv
#> 220 4 28        suv
#> 221 5 28        suv
#> 222 6 28        suv
#> 223 7 28        suv
#> 224 8 28        suv
#> 225 1 29        suv
#> 226 2 29        suv
#> 227 3 29        suv
#> 228 4 29        suv
#> 229 5 29        suv
#> 230 6 29        suv
#> 231 7 29        suv
#> 232 8 29        suv
#> 233 1 30        suv
#> 234 2 30        suv
#> 235 3 30       <NA>
#> 236 4 30       <NA>
#> 237 5 30       <NA>
#> 238 6 30       <NA>
#> 239 7 30       <NA>
#> 240 8 30       <NA>

waffle_iron(mpg, aes_d(group = class), 0.75)

#>     y   x      group
#> 1   1   1    2seater
#> 2   1   2    2seater
#> 3   1   3    2seater
#> 4   1   4    2seater
#> 5   1   5    2seater
#> 6   1   6    compact
#> 7   1   7    compact
#> 8   1   8    compact
#> 9   1   9    compact
#> 10  1  10    compact
#> 11  1  11    compact
#> 12  1  12    compact
#> 13  1  13    compact
#> 14  1  14    compact
#> 15  1  15    compact
#> 16  1  16    compact
#> 17  1  17    compact
#> 18  1  18    compact
#> 19  1  19    compact
#> 20  1  20    compact
#> 21  1  21    compact
#> 22  1  22    compact
#> 23  1  23    compact
#> 24  1  24    compact
#> 25  1  25    compact
#> 26  1  26    compact
#> 27  1  27    compact
#> 28  1  28    compact
#> 29  1  29    compact
#> 30  1  30    compact
#> 31  1  31    compact
#> 32  1  32    compact
#> 33  1  33    compact
#> 34  1  34    compact
#> 35  1  35    compact
#> 36  1  36    compact
#> 37  1  37    compact
#> 38  1  38    compact
#> 39  1  39    compact
#> 40  1  40    compact
#> 41  1  41    compact
#> 42  1  42    compact
#> 43  1  43    compact
#> 44  1  44    compact
#> 45  1  45    compact
#> 46  1  46    compact
#> 47  1  47    compact
#> 48  1  48    compact
#> 49  1  49    compact
#> 50  1  50    compact
#> 51  1  51    compact
#> 52  1  52    compact
#> 53  1  53    midsize
#> 54  1  54    midsize
#> 55  1  55    midsize
#> 56  1  56    midsize
#> 57  1  57    midsize
#> 58  1  58    midsize
#> 59  1  59    midsize
#> 60  1  60    midsize
#> 61  1  61    midsize
#> 62  1  62    midsize
#> 63  1  63    midsize
#> 64  1  64    midsize
#> 65  1  65    midsize
#> 66  1  66    midsize
#> 67  1  67    midsize
#> 68  1  68    midsize
#> 69  1  69    midsize
#> 70  1  70    midsize
#> 71  1  71    midsize
#> 72  1  72    midsize
#> 73  1  73    midsize
#> 74  1  74    midsize
#> 75  1  75    midsize
#> 76  1  76    midsize
#> 77  1  77    midsize
#> 78  1  78    midsize
#> 79  1  79    midsize
#> 80  1  80    midsize
#> 81  1  81    midsize
#> 82  1  82    midsize
#> 83  1  83    midsize
#> 84  1  84    midsize
#> 85  1  85    midsize
#> 86  1  86    midsize
#> 87  1  87    midsize
#> 88  1  88    midsize
#> 89  1  89    midsize
#> 90  1  90    midsize
#> 91  1  91    midsize
#> 92  1  92    midsize
#> 93  1  93    midsize
#> 94  1  94    minivan
#> 95  1  95    minivan
#> 96  1  96    minivan
#> 97  1  97    minivan
#> 98  1  98    minivan
#> 99  1  99    minivan
#> 100 1 100    minivan
#> 101 1 101    minivan
#> 102 1 102    minivan
#> 103 1 103    minivan
#> 104 1 104    minivan
#> 105 1 105     pickup
#> 106 1 106     pickup
#> 107 1 107     pickup
#> 108 1 108     pickup
#> 109 1 109     pickup
#> 110 1 110     pickup
#> 111 1 111     pickup
#> 112 1 112     pickup
#> 113 1 113     pickup
#> 114 1 114     pickup
#> 115 1 115     pickup
#> 116 1 116     pickup
#> 117 1 117     pickup
#> 118 1 118     pickup
#> 119 1 119     pickup
#> 120 1 120     pickup
#> 121 1 121     pickup
#> 122 1 122     pickup
#> 123 1 123     pickup
#> 124 1 124     pickup
#> 125 1 125     pickup
#> 126 1 126     pickup
#> 127 1 127     pickup
#> 128 1 128     pickup
#> 129 1 129     pickup
#> 130 1 130     pickup
#> 131 1 131     pickup
#> 132 1 132     pickup
#> 133 1 133     pickup
#> 134 1 134     pickup
#> 135 1 135     pickup
#> 136 1 136     pickup
#> 137 1 137     pickup
#> 138 1 138 subcompact
#> 139 1 139 subcompact
#> 140 1 140 subcompact
#> 141 1 141 subcompact
#> 142 1 142 subcompact
#> 143 1 143 subcompact
#> 144 1 144 subcompact
#> 145 1 145 subcompact
#> 146 1 146 subcompact
#> 147 1 147 subcompact
#> 148 1 148 subcompact
#> 149 1 149 subcompact
#> 150 1 150 subcompact
#> 151 1 151 subcompact
#> 152 1 152 subcompact
#> 153 1 153 subcompact
#> 154 1 154 subcompact
#> 155 1 155 subcompact
#> 156 1 156 subcompact
#> 157 1 157 subcompact
#> 158 1 158 subcompact
#> 159 1 159 subcompact
#> 160 1 160 subcompact
#> 161 1 161 subcompact
#> 162 1 162 subcompact
#> 163 1 163 subcompact
#> 164 1 164 subcompact
#> 165 1 165 subcompact
#> 166 1 166 subcompact
#> 167 1 167 subcompact
#> 168 1 168 subcompact
#> 169 1 169 subcompact
#> 170 1 170 subcompact
#> 171 1 171 subcompact
#> 172 1 172 subcompact
#> 173 1 173        suv
#> 174 1 174        suv
#> 175 1 175        suv
#> 176 1 176        suv
#> 177 1 177        suv
#> 178 1 178        suv
#> 179 1 179        suv
#> 180 1 180        suv
#> 181 1 181        suv
#> 182 1 182        suv
#> 183 1 183        suv
#> 184 1 184        suv
#> 185 1 185        suv
#> 186 1 186        suv
#> 187 1 187        suv
#> 188 1 188        suv
#> 189 1 189        suv
#> 190 1 190        suv
#> 191 1 191        suv
#> 192 1 192        suv
#> 193 1 193        suv
#> 194 1 194        suv
#> 195 1 195        suv
#> 196 1 196        suv
#> 197 1 197        suv
#> 198 1 198        suv
#> 199 1 199        suv
#> 200 1 200        suv
#> 201 1 201        suv
#> 202 1 202        suv
#> 203 1 203        suv
#> 204 1 204        suv
#> 205 1 205        suv
#> 206 1 206        suv
#> 207 1 207        suv
#> 208 1 208        suv
#> 209 1 209        suv
#> 210 1 210        suv
#> 211 1 211        suv
#> 212 1 212        suv
#> 213 1 213        suv
#> 214 1 214        suv
#> 215 1 215        suv
#> 216 1 216        suv
#> 217 1 217        suv
#> 218 1 218        suv
#> 219 1 219        suv
#> 220 1 220        suv
#> 221 1 221        suv
#> 222 1 222        suv
#> 223 1 223        suv
#> 224 1 224        suv
#> 225 1 225        suv
#> 226 1 226        suv
#> 227 1 227        suv
#> 228 1 228        suv
#> 229 1 229        suv
#> 230 1 230        suv
#> 231 1 231        suv
#> 232 1 232        suv
#> 233 1 233        suv
#> 234 1 234        suv
#> 235 1 235       <NA>
#> 236 1 236       <NA>
#> 237 1 237       <NA>
#> 238 1 238       <NA>
#> 239 1 239       <NA>
#> 240 1 240       <NA>
#> 241 1 241       <NA>
#> 242 1 242       <NA>
#> 243 1 243       <NA>
#> 244 1 244       <NA>
#> 245 1 245       <NA>
#> 246 1 246       <NA>
#> 247 1 247       <NA>
#> 248 1 248       <NA>
#> 249 1 249       <NA>
#> 250 1 250       <NA>
#> 251 1 251       <NA>
#> 252 1 252       <NA>
#> 253 1 253       <NA>
#> 254 1 254       <NA>
#> 255 1 255       <NA>
#> 256 1 256       <NA>
#> 257 1 257       <NA>
#> 258 1 258       <NA>
#> 259 1 259       <NA>
#> 260 1 260       <NA>
#> 261 1 261       <NA>
#> 262 1 262       <NA>
#> 263 1 263       <NA>
#> 264 1 264       <NA>
#> 265 1 265       <NA>
#> 266 1 266       <NA>
#> 267 1 267       <NA>
#> 268 1 268       <NA>
#> 269 1 269       <NA>
#> 270 1 270       <NA>
#> 271 1 271       <NA>
#> 272 1 272       <NA>
#> 273 1 273       <NA>
#> 274 1 274       <NA>
#> 275 1 275       <NA>
#> 276 1 276       <NA>
#> 277 1 277       <NA>
#> 278 1 278       <NA>
#> 279 1 279       <NA>
#> 280 1 280       <NA>
#> 281 1 281       <NA>
#> 282 1 282       <NA>
#> 283 1 283       <NA>
#> 284 1 284       <NA>
#> 285 1 285       <NA>
#> 286 1 286       <NA>
#> 287 1 287       <NA>
#> 288 1 288       <NA>
#> 289 1 289       <NA>
#> 290 1 290       <NA>
#> 291 1 291       <NA>
#> 292 1 292       <NA>
#> 293 1 293       <NA>
#> 294 1 294       <NA>
#> 295 1 295       <NA>
#> 296 1 296       <NA>
#> 297 1 297       <NA>
#> 298 1 298       <NA>
#> 299 1 299       <NA>
#> 300 1 300       <NA>
#> 301 1 301       <NA>
#> 302 1 302       <NA>
#> 303 1 303       <NA>
#> 304 1 304       <NA>
#> 305 1 305       <NA>
#> 306 1 306       <NA>
#> 307 1 307       <NA>
#> 308 1 308       <NA>
#> 309 1 309       <NA>
#> 310 1 310       <NA>
#> 311 1 311       <NA>
#> 312 1 312       <NA>